+++ .changeset/3755-consolidate-init-tests.md
@@ -1,0 +1,7 @@
+---
+type: Changed
+pr: 3755
+---
+Internal test refactor: consolidates the Init Command Module test cluster from 7 files to 5.
+
+<!-- docs-exempt: internal test refactor only — no user-facing surface changed -->
