+++ .gitignore
@@ -70,3 +70,11 @@ vendor/
 tmp/
 .worktrees
 .envrc
+
+# MemPalace per-project files (issue #185)
+mempalace.yaml
+entities.json
+
+# Local scratch + Claude-test artifacts
+.scratch/
+claude-test-command.md
